requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
pytz==2024.2
google-generativeai==0.8.3
requests==2.32.3
